# ARM64 assembly optimized vector operations