--- a/discmath/oct_arm64.s
+++ b/discmath/oct_arm64.s
@@ -112,7 +112,6 @@ loop16:
     SUBS   $1, R3
     BNE    loop16
 
-// ---- tail <16: 8/4/2/1 via TBZ (no NEON) ----
 tail:
     TBZ    $3, R2, chk4
     MOVD   0(R0), R3
@@ -146,12 +145,8 @@ chk1:
 done:
     RET
 
-DATA  ·mask0f+0(SB)/8, $0x0f0f0f0f0f0f0f0f
-DATA  ·mask0f+8(SB)/8, $0x0f0f0f0f0f0f0f0f
-GLOBL ·mask0f(SB), RODATA, $16
-
 // func OctVecMul(vector []byte, multiplier uint8)
-TEXT ·OctVecMul(SB), $0-32
+TEXT ·OctVecMul(SB), NOSPLIT|NOFRAME, $0-32
     MOVD   vector+0(FP), R0          // R0 = &vector[0]
     MOVD   vector_len+8(FP), R1      // R1 = len(vector)
     MOVBU  multiplier+24(FP), R2     // R2 = multiplier
@@ -165,8 +160,7 @@ TEXT ·OctVecMul(SB), $0-32
     VLD1   (R3), [V2.B16]            // V2 = rowLo
     VLD1   (R4), [V3.B16]            // V3 = rowHi
 
-    MOVD   $·mask0f(SB), R6
-    VLD1   (R6), [V1.B16]            // V1 = 0x0f0f..0f0f
+    VMOVI  $0x0f, V1.B16             // V1 = 0x0F mask
 
     LSR     $4, R1, R5
     CBZ     R5, tail
@@ -200,5 +194,77 @@ tail:
     ADD     $1, R0
     SUBS    $1, R1
     BNE     tail
+done:
+    RET
+
+// func OctVecMulAdd(x, y []byte, multiplier uint8)
+TEXT ·OctVecMulAdd(SB), NOSPLIT|NOFRAME, $0-56
+    MOVD   x_base+0(FP),  R0         // R0 = &x[0]
+    MOVD   x_len+8(FP),   R2         // R2 = len(x)
+    MOVD   y_base+24(FP), R1         // R1 = &y[0]
+    MOVBU  multiplier+48(FP), R5     // R5 = multiplier
+
+    MOVD   $·_OctMulLo(SB), R3       // R3 = base of OctMulLo
+    MOVD   $·_OctMulHi(SB), R4       // R4 = base of OctMulHi
+    LSL    $4, R5, R6                // R6 = multiplier * 16
+    ADD    R6, R3, R3                // R3 = &OctMulLo[multiplier*16]
+    ADD    R6, R4, R4                // R4 = &OctMulHi[multiplier*16]
+
+    VLD1   (R3), [V2.B16]            // V2 = rowLo
+    VLD1   (R4), [V3.B16]            // V3 = rowHi
+    VMOVI  $0x0f, V15.B16            // V15 = 0x0F mask
+
+before32:
+    LSR    $5, R2, R6                // R6 = n / 32
+    CBZ    R6, tail
+
+loop32:
+    VLD1.P 32(R1), [V0.B16, V1.B16]  // V0=y0, V1=y1; y += 32
+    VLD1   (R0),   [V4.B16, V5.B16]  // V4=x0, V5=x1
+
+    VAND   V15.B16, V0.B16, V6.B16   // V6 = y0 & 0x0F
+    VUSHR  $4,      V0.B16, V0.B16   // V0 = y0 >> 4
+    VAND   V15.B16, V1.B16, V7.B16   // V7 = y1 & 0x0F
+    VUSHR  $4,      V1.B16, V1.B16   // V1 = y1 >> 4
+
+    VTBL   V6.B16, [V2.B16], V6.B16  // V6 = rowLo[low0]
+    VTBL   V7.B16, [V2.B16], V7.B16  // V7 = rowLo[low1]
+    VTBL   V0.B16, [V3.B16], V0.B16  // V0 = rowHi[high0]
+    VTBL   V1.B16, [V3.B16], V1.B16  // V1 = rowHi[high1]
+
+    VEOR   V6.B16, V0.B16, V6.B16    // V6 = mul(y0)
+    VEOR   V7.B16, V1.B16, V7.B16    // V7 = mul(y1)
+
+    VEOR   V6.B16, V4.B16, V4.B16    // x0 ^= mul(y0)
+    VEOR   V7.B16, V5.B16, V5.B16    // x1 ^= mul(y1)
+    VST1.P [V4.B16, V5.B16], 32(R0)  // store; x += 32
+
+    SUBS   $1, R6
+    BNE    loop32
+after32:
+    AND    $31, R2, R2
+    CBZ    R2, done
+
+tail:
+    MOVBU  0(R1), R8                // R8 = y[i]
+    AND    $0x0F, R8, R9            // R9 = y[i] & 0x0F
+    ADD    R3, R9, R9
+    MOVBU  0(R9), R9                // R9 = rowLo[lo]
+
+    LSR    $4, R8, R8               // R8 = y[i] >> 4
+    ADD    R4, R8, R8
+    MOVBU  0(R8), R8                // R8 = rowHi[hi]
+
+    EOR    R8, R9, R9               // R9 = rowLo[lo] ^ rowHi[hi] = mul(y[i])
+
+    MOVBU  0(R0), R10               // R10 = x[i]
+    EOR    R9, R10, R10             // R10 = x[i] ^ mul(y[i])
+    MOVB   R10, 0(R0)               // store result
+
+    ADD    $1, R0                   // x++
+    ADD    $1, R1                   // y++
+    SUBS   $1, R2
+    BNE    tail
+
 done:
     RET