--- a/discmath/oct_arm64.s
+++ b/discmath/oct_arm64.s
@@ -1,4 +1,4 @@
-//go:build arm64
+//go:build arm64 && !purego
 #include "textflag.h"
 
 // func OctVecAdd(x, y []byte)