--- a/discmath/oct_generic.go
+++ b/discmath/oct_generic.go
@@ -1,4 +1,4 @@
-//go:build !amd64 && !arm64
+//go:build (!amd64 && !arm64) || purego
 
 package discmath