… discmath/oct_amd64.s