Benchmarks:
```
Apple M4 Max - arm64 assembly:
Benchmark_EncodeDecodeFuzz-16         22200       53258 ns/op    354866 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-16         20676       53278 ns/op    354866 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-16         22441       53261 ns/op    354866 B/op       169 allocs/op

Apple M4 Max - purego:
Benchmark_EncodeDecodeFuzz-16          8996      122385 ns/op    354866 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-16          9306      124842 ns/op    354865 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-16          9254      121512 ns/op    354866 B/op       169 allocs/op

Cortex-A76 (Raspberry Pi 5) - arm64 assembly:
Benchmark_EncodeDecodeFuzz-4         3987      313808 ns/op    354834 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-4         3958      339169 ns/op    354834 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-4         3902      316628 ns/op    354833 B/op       169 allocs/op

Cortex-A76 (Raspberry Pi 5) - purego:
Benchmark_EncodeDecodeFuzz-4         1974      609037 ns/op    354836 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-4         2042      573302 ns/op    354836 B/op       169 allocs/op
Benchmark_EncodeDecodeFuzz-4         1824      636907 ns/op    354836 B/op       169 allocs/op
```